annotation-like objects axis reference mapping
to a subplot on a new figure. This will allow us to copy the annotations
to a new figure with more axes as part of a px.overlay command.

This just needs to be generalized to work with shapes and images but the
workings can be seen in proto/px_overlay/map_axis_pair_example.py.