Added backslashes to descriptions with '[][]'
Changes were made to line 2037, in the case where a node is valType info_array in the JSON with dimension: 2 (eg. groups in _node.py line 137).
This fixes the issue where markdown interprets it as a hyperlink. This line adds '[][]' to the description after the description passed in the JSON file as part of the validation process.
Now, markdown correctly interprets the brackets as consecutive brackets instead.